Types and Functions

Types and Functions

Operations

LandauGinzburgCategories.Operations.:⨷ — Function.

A⨷B

Graded tensor product of ℤ/2 graded block matrices. The result consists of even/odd blocks just like the input and the operation is associative.

Graded means that we introduce Koszul signs. Writing $A = a^i_j e_i e^j$ and $B = b^k_ℓ f_k f^ℓ$ for some (co)basis vectors, we represent the tensor product as

$A⨷B = (-1)^{|j|(|k|+|ℓ|)} p^i_j q^k_ℓ e_i f_k f^ℓ e^j$

where the sign comes from commuting $e^j$ with $f_k f^ℓ$. This ensures that, upon matrix multiplication between two such tensor products, the contractions between $e^j$ and $e_j$, and between $f^ℓ$ and $f_ℓ$, are already adjacent and do not introduce more signs.

Mapping the multi-indices $(i,k)$ to rows and $(ℓ,j)$ to columns requires some care to ensure that we end up with even/odd blocks. A comment in the code explains how this is done. We first interlace the rows/columns to get alternating grades in rows/columns (to_alternating_grades and from_alternating_grades). Then we need only to reverse some row/column orders. Finally, we de-interlace to get even/odd blocks. (I wonder if there is a way to avoid making the detour through the alternating signs representation, but it seems hard to maintain associativity.)

source

LandauGinzburgCategories.Operations.:⨶ — Function.

A⨶B

Tensor product of matrix factorizations.

This is the composition operation in the bicategory or Landau Ginzburg models.

source

Q,ϵ = ⨶(A,B,W,vars...)

Finite-rank homotopy representation of $A⨶B$, where we remove the variables vars from the result.

We return a matrix factorization $Q$ together with an idempotent ϵ representing the direct summand of $Q$ that represents $A⨶B$.

See the pushforward paper by Dykerhoff&Murfet.

source

LandauGinzburgCategories.QuasiHomogeneous.centralcharge — Function.

c = centralcharge(f, vars...)

Return the central charge of f with respect to the variables vars.

This is the value of the expression

$\sum_{i=1}^n 1 - q_i$

where $q_i$ is the grading of the $i$th variable under a (ℚ-valued) grading for which f is homogeneous of degree 2.

source

LandauGinzburgCategories.Operations.dual — Function.

Q′ = dual(Q)

Return the dual (i.e. the adjoint) matrix factorization of Q.

If Q factors the potential f, then this is a matrix factorization that factors -f.

source

LandauGinzburgCategories.QuasiHomogeneous.find_quasihomogeneous_degrees — Function.

gradings = find_quasihomogeneous_degrees(f, vars...)

Find the assignment of gradings to the variables vars such that f is a quasi-homogeneous polynomial with respect to these gradings, if such an assignment exists, and return it. Otherwise, raise an exception.

Example

julia> find_quasihomogeneous_degrees(x^4*y + x*y^9, :x, :y)
(x = 8, y = 3)

source

LandauGinzburgCategories.Operations.fuse — Function.

docstring goes here

source

LandauGinzburgCategories.Operations.fuse_abstract — Function.

docstring goes here

source

LandauGinzburgCategories.Operations.unit_matrix_factorization — Function.

unit_matrix_factorization(f; source_to_target...)

A ℤ/2-graded matrix that squares to f(;source_to_target...) - f times the identity matrix.

The source for this formulation is

Adjunctions and defects in Landau-Ginzburg models, Nils Carqueville and Daniel Murfet

source

Library

LandauGinzburgCategories.Library.orbifold_equivalence — Function.

orbifold_equivalence(f, g)

Return a matrix representing an orbifold equivalence between f and g, if one is available in the library. Return false if it is known that f and g are not orbifold equivalent. Return missing if this is not known by this library.

source

LandauGinzburgCategories.Library.Potential — Type.

Potential{NumVars}

A type representing a potential through a classification scheme. For example, the potential $x^4 - y^2$ is called $A_3$ through the ADE classification of unimodular potentials, and it is represented in Julia by the type TwoVariables.A₃. This is a subtype of Potential{2}.

source

Internal functions

LandauGinzburgCategories.Operations.flatten_blocks — Function.

M = flatten_blocks(X)

Construct a matrix from a matrix of matrices by concatenating them horizontally and vertically.

source

LandauGinzburgCategories.Operations.block_diagonalization — Function.

D, A = block_diagonalization(X)

Decompose the matrix factorization X into a direct sum of irreducible matrix factorizations, and represent this direct sum as a block-diagonal matrix factorization D such that $A^{-1} D A = X$.

NOTE: for now, this function only returns D, and not yet A!

source